style: code formatting as per review
