--- src/routers/profile.router.js
@@ -1,7 +1,0 @@
-import express from 'express';
-import profileController from '../controllers/profile.controller';
-
-const router = express.Router();
-router.post('/:handle/profile', profileController.profile);
-
-export default router;
