+++ src/utils/axios/gitHubProxy.js
@@ -11,6 +11,7 @@ class GitHubProxy {
   }
 
   static async getHandler(handle) {
+    // console.log(handle);
     const handler = await axios.get(
       `${config.get().gitHub.api.baseUrl}/users/${handle}`,
     );
