+++ src/utils/logger/logger.js
@@ -4,9 +4,9 @@ import { createLogger, format, transports } from 'winston';
 const year = moment().format('Y');
 const month = moment().format('MMMM');
 const day = moment().format('DD');
-const repoLogger = createLogger({
+const logger = createLogger({
   transports: new transports.File({
-    filename: `logs/repository/${year}/${month}/${day}/log_${year}_${month}_${day}.log`,
+    filename: `logs/${year}/${month}/${day}/log_${year}_${month}_${day}.log`,
     format: format.combine(
       format.errors({ stack: true }),
       format.timestamp({ format: 'MMM-DD-YYYY HH:mm:ss' }),
@@ -21,21 +21,4 @@ const repoLogger = createLogger({
   }),
 });
 
-const profileLogger = createLogger({
-  transports: new transports.File({
-    filename: `logs/profile/${year}/${month}/${day}/log_${year}_${month}_${day}.log`,
-    format: format.combine(
-      format.errors({ stack: true }),
-      format.timestamp({ format: 'MMM-DD-YYYY HH:mm:ss' }),
-      format.printf((info) => {
-        if (typeof info.message === 'object') {
-          info.message = JSON.stringify(info.message, null, 3);
-        }
-
-        return `${info.level}: ${[info.timestamp]}: ${info.message}`;
-      }),
-    ),
-  }),
-});
-
-export { repoLogger, profileLogger };
+export default logger;
